fix: when Monthly day was set to a numbe higher than 28, it caused issues in February
For example, Task was set to 29'th day.
On february 28, it checked if there is a future task already created, and it was not.
So the task log was created with 28/02 as the date

Next time user opened tasks, the same check was done. And again it found future task log not created.
And it tried to create again a task log with date 28/02, but that already existed and failed.